+++ Cargo.toml
@@ -96,13 +96,8 @@ async-compat = "0.2.1"
 clap = { version = "3.1.12", optional = true, features = ["derive"] }
 
 wasm-bindgen-futures = "0.4.30"
-serde = { version = "1.0.80", features = ["derive"] }
-serde_derive = "^1.0.59"
 futures = "0.3"
 range-reader = "0.1"
-wasm-rs-async-executor = { version = "0.9.0", features = ["cooperative-browser", "debug", "requestIdleCallback"] }
-
-tokio = { version = "1.1", features = ["macros", "time", "sync"] }
 
 [dependencies.web-sys]
 version = "0.3.4"
