+++ src/arrow1/mod.rs
@@ -1,12 +1,6 @@
 #[cfg(feature = "reader")]
 pub mod reader;
 
-#[cfg(feature = "reader")]
-pub mod reader_async;
-
-#[cfg(feature = "reader")]
-pub mod ranged_reader;
-
 pub mod wasm;
 
 #[cfg(feature = "writer")]
