--- src/async_testing.rs
@@ -1,127 +1,0 @@
-use serde::{Deserialize, Serialize};
-use wasm_bindgen::prelude::*;
-use wasm_bindgen::JsCast;
-use wasm_bindgen_futures::JsFuture;
-use web_sys::{Request, RequestInit, RequestMode, Response};
-
-/// A struct to hold some data from the github Branch API.
-///
-/// Note how we don't have to define every member -- serde will ignore extra
-/// data when deserializing
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Branch {
-    pub name: String,
-    pub commit: Commit,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Commit {
-    pub sha: String,
-    pub commit: CommitDetails,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct CommitDetails {
-    pub author: Signature,
-    pub committer: Signature,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Signature {
-    pub name: String,
-    pub email: String,
-}
-
-// A macro to provide `println!(..)`-style syntax for `console.log` logging.
-macro_rules! log {
-    ( $( $t:tt )* ) => {
-        web_sys::console::log_1(&format!( $( $t )* ).into());
-    }
-}
-
-#[wasm_bindgen]
-pub async fn getContentLength(url: String) -> Result<(), JsValue> {
-    let mut opts = RequestInit::new();
-    opts.method("HEAD");
-    opts.mode(RequestMode::Cors);
-
-    let request = Request::new_with_str_and_init(&url, &opts)?;
-
-    request
-        .headers()
-        .set("Accept", "application/vnd.github.v3+json")?;
-
-    let window = web_sys::window().unwrap();
-    let resp_value = JsFuture::from(window.fetch_with_request(&request)).await?;
-    let resp: Response = resp_value.dyn_into().unwrap();
-    let length = resp.headers().get("content-length")?;
-    let a = length.unwrap();
-    let lengthInt = a.parse::<u32>().unwrap();
-
-    log!("{lengthInt:?}");
-    // log!("{length:?}");
-    log!("{resp:?}");
-
-    Ok(())
-}
-
-#[wasm_bindgen]
-pub async fn makeRangeRequest(url: String) -> Result<(), JsValue> {
-    let mut opts = RequestInit::new();
-    opts.method("GET");
-    opts.mode(RequestMode::Cors);
-
-    let request = Request::new_with_str_and_init(&url, &opts)?;
-
-    request
-        .headers()
-        .set("Range", "bytes=0-1023")?;
-
-    let window = web_sys::window().unwrap();
-    let resp_value = JsFuture::from(window.fetch_with_request(&request)).await?;
-    let resp: Response = resp_value.dyn_into().unwrap();
-    let buffer = JsFuture::from(resp.array_buffer().ok().unwrap()).await?;
-
-    log!("{resp:?}");
-    log!("{buffer:?}");
-    // let length = resp.headers().get("content-length")?;
-    // let a = length.unwrap();
-    // let lengthInt = a.parse::<u32>().unwrap();
-
-    // log!("{lengthInt:?}");
-    // // log!("{length:?}");
-
-    Ok(())
-}
-
-
-#[wasm_bindgen]
-pub async fn run(repo: String) -> Result<JsValue, JsValue> {
-    let mut opts = RequestInit::new();
-    opts.method("GET");
-    opts.mode(RequestMode::Cors);
-
-    let url = format!("https://api.github.com/repos/{}/branches/main", repo);
-
-    let request = Request::new_with_str_and_init(&url, &opts)?;
-
-    request
-        .headers()
-        .set("Accept", "application/vnd.github.v3+json")?;
-
-    let window = web_sys::window().unwrap();
-    let resp_value = JsFuture::from(window.fetch_with_request(&request)).await?;
-
-    // `resp_value` is a `Response` object.
-    assert!(resp_value.is_instance_of::<Response>());
-    let resp: Response = resp_value.dyn_into().unwrap();
-
-    // Convert this other `Promise` into a rust `Future`.
-    let json = JsFuture::from(resp.json()?).await?;
-
-    // Use serde to parse the JSON into a struct.
-    let branch_info: Branch = json.into_serde().unwrap();
-
-    // Send the `Branch` struct back to JS as an `Object`.
-    Ok(JsValue::from_serde(&branch_info).unwrap())
-}
